fix(agent): use Responses API for OpenRouter OpenAI models (#2501)
* fix(agent): use Responses API for OpenRouter

* fix(agent): preserve Responses request options and errors

* test(agent): remove OpenRouter Responses coverage

* fix(agent): scope OpenRouter Responses to OpenAI models

* docs(agent): clarify OpenRouter BYOK fix

* fix(agent): preserve direct OpenAI temperature behavior
