+++ libraries/typescript/.changeset/calm-router-responses.md
@@ -1,0 +1,5 @@
+---
+"@mcp-use/agent": patch
+---
+
+Fix OpenRouter BYOK requests to OpenAI models returning neither a model response nor an actionable provider error by routing those models through the Responses API.
