+++ libraries/typescript/packages/agent/src/llm/providers/openai-chat-completions.ts
@@ -11,43 +11,9 @@ import type {
   ProviderTool,
 } from "../types.js";
 import { tokenUsageFromRecord } from "../usage.js";
+import { throwLlmRequestError } from "./openai-shared.js";
 
-/** HTTP error returned by an LLM provider request. */
-export class LlmRequestError extends Error {
-  /** HTTP response status. */
-  readonly status: number;
-  /** Parsed JSON response body, or the raw response text. */
-  readonly body?: unknown;
-
-  /**
-   * @param status - HTTP response status.
-   * @param message - Error message.
-   * @param body - Parsed or raw provider response body.
-   */
-  constructor(status: number, message: string, body?: unknown) {
-    super(message);
-    this.name = "LlmRequestError";
-    this.status = status;
-    this.body = body;
-  }
-}
-
-async function throwLlmRequestError(res: Response): Promise<never> {
-  const text = await res.text().catch(() => "");
-  let body: unknown = text;
-  if (text) {
-    try {
-      body = JSON.parse(text);
-    } catch {
-      // keep raw text
-    }
-  }
-  throw new LlmRequestError(
-    res.status,
-    `OpenAI request failed (${res.status} ${res.statusText}): ${text}`,
-    body
-  );
-}
+export { LlmRequestError } from "./openai-shared.js";
 
 interface ChatParams {
   config: ProviderConfig;
